decouple glightbox and mathjax
