+++ front-react/src/hooks/use-conversation-html-glightbox.ts
@@ -1,5 +1,76 @@
 import { useCallback, useEffect, useRef } from "react";
-import { type GlightboxInstance, preloadGlightbox } from "./conversation-html-assets";
+
+type GlightboxInstance = {
+    destroy?: () => void;
+};
+
+declare global {
+    interface Window {
+        GLightbox?: (options?: { selector?: string }) => GlightboxInstance;
+    }
+}
+
+let glightboxPromise: Promise<void> | null = null;
+
+function loadStylesheet(id: string, href: string) {
+    if (document.getElementById(id)) {
+        return;
+    }
+
+    const link = document.createElement("link");
+    link.id = id;
+    link.rel = "stylesheet";
+    link.href = href;
+    document.head.appendChild(link);
+}
+
+function loadScript(id: string, src: string): Promise<void> {
+    return new Promise((resolve, reject) => {
+        const existingScript = document.getElementById(id) as HTMLScriptElement | null;
+
+        if (existingScript?.dataset.loaded === "true") {
+            resolve();
+            return;
+        }
+
+        const script = existingScript ?? document.createElement("script");
+
+        const handleLoad = () => {
+            script.dataset.loaded = "true";
+            resolve();
+        };
+
+        const handleError = () => {
+            reject(new Error(`Failed to load script: ${src}`));
+        };
+
+        script.addEventListener("load", handleLoad, { once: true });
+        script.addEventListener("error", handleError, { once: true });
+
+        if (!existingScript) {
+            script.id = id;
+            script.src = src;
+            script.async = true;
+            document.head.appendChild(script);
+        }
+    });
+}
+
+function preloadGlightbox(): Promise<void> {
+    glightboxPromise ??= (async () => {
+        loadStylesheet(
+            "conversation-panel-glightbox-style",
+            "https://cdn.jsdelivr.net/npm/glightbox/dist/css/glightbox.min.css"
+        );
+
+        await loadScript(
+            "conversation-panel-glightbox-script",
+            "https://cdn.jsdelivr.net/npm/glightbox/dist/js/glightbox.min.js"
+        );
+    })();
+
+    return glightboxPromise;
+}
 
 type UseConversationHtmlGlightboxOptions = {
     conversationId: string | undefined;
